int to long long

+++ Answered/D.cpp
@@ -1,10 +1,10 @@
 #include<iostream>
 using namespace std;
-int n, ans = 0;
+long long n, ans = 0;
 
 int main() {
 	cin >> n;
-	for (int i = 1; i <= n; i++)
+	for (long long i = 1; i <= n; i++)
 	{
 		ans += i;
 	}
